feat : changed the view to make the main pannel unscrollable in most cases (#87)
